Only use final speech transcript
The SpeechRecognition API provides an interim transcript that includes
guesses made by the transcription algorithm for words that are currently
being spoken. This means the output can change multiple times for a
particular segment of speech. Change the Game component to only render
the final transcript, which happens when the browser has maximum
confidence in its guess.