Conditionally import chess.js
chess.js exports an object

    { Chess: Chess }

on CommonJS environments but simply exports the Chess constructor on
AMD environments (https://github.com/jhlywa/chess.js/issues/196). This
causes a compilation error in either the browser or in Node if you run
the same code in both. Which one fails depends on how you do the import.
This commit does the import conditionally depending on the environment.